+++ pkg/test/context.go
@@ -24,11 +24,7 @@ import (
 	"k8s.io/client-go/restmapper"
 )
 
-type TestCtx struct { //nolint:golint
-	// todo(camilamacedo86): The no lint here is for type name will be used as test.TestCtx by other packages, and
-	//  that stutters; consider calling this Ctx (golint)
-	// However, was decided to not move forward with it now in order to not introduce breakchanges with the task to
-	// add the linter. We should to do it after.
+type Context struct {
 	id         string
 	cleanupFns []cleanupFn
 	namespace  string
@@ -41,43 +37,57 @@ type TestCtx struct { //nolint:golint
 	skipCleanupOnError bool
 }
 
+// todo(camilamacedo86): Remove the following line just added for we are able to deprecated TestCtx
+// need to be done before: 1.0.0
+
+// Deprecated: Context exists for historical compatibility
+// Yoou should use Context instead of.
+type TestCtx = Context //nolint:golint
+
 type CleanupOptions struct {
-	TestContext   *TestCtx
+	TestContext   *Context
 	Timeout       time.Duration
 	RetryInterval time.Duration
 }
 
 type cleanupFn func() error
 
-func (f *Framework) newTestCtx(t *testing.T) *TestCtx {
-	// TestCtx is used among others for namespace names where '/' is forbidden and must be 63 characters or less
+func (f *Framework) newContext(t *testing.T) *Context {
+
+	// Context is used among others for namespace names where '/' is forbidden and must be 63 characters or less
 	id := "osdk-e2e-" + uuid.New()
 
 	var namespace string
 	if f.singleNamespaceMode {
 		namespace = f.Namespace
 	}
-	return &TestCtx{
-		id:                 id,
-		t:                  t,
-		namespace:          namespace,
-		namespacedManPath:  *f.NamespacedManPath,
-		client:             f.Client,
-		kubeclient:         f.KubeClient,
-		restMapper:         f.restMapper,
-		skipCleanupOnError: f.skipCleanupOnError,
+	return &Context{
+		id:                id,
+		t:                 t,
+		namespace:         namespace,
+		namespacedManPath: *f.NamespacedManPath,
+		client:            f.Client,
+		kubeclient:        f.KubeClient,
+		restMapper:        f.restMapper,
 	}
 }
 
-func NewTestCtx(t *testing.T) *TestCtx {
-	return Global.newTestCtx(t)
+// Deprecated: NewTestCtx exists for historical compatibility
+// You should use NewContext instead of.
+// todo(camilamacedo86): remove before 1.0.0
+func NewTestCtx(t *testing.T) *Context {
+	return Global.newContext(t)
 }
 
-func (ctx *TestCtx) GetID() string {
+func NewContext(t *testing.T) *Context {
+	return Global.newContext(t)
+}
+
+func (ctx *Context) GetID() string {
 	return ctx.id
 }
 
-func (ctx *TestCtx) Cleanup() {
+func (ctx *Context) Cleanup() {
 	if ctx.t != nil {
 		// The cleanup function will be skipped
 		if ctx.t.Failed() && ctx.skipCleanupOnError {
@@ -103,6 +113,6 @@ func (ctx *TestCtx) Cleanup() {
 	}
 }
 
-func (ctx *TestCtx) AddCleanupFn(fn cleanupFn) {
+func (ctx *Context) AddCleanupFn(fn cleanupFn) {
 	ctx.cleanupFns = append(ctx.cleanupFns, fn)
 }