+++ test/e2e/tls_util_test.go
@@ -82,7 +82,7 @@ func init() {
 // and CA TLS assets exist in the k8s cluster for a given cr,
 // the GenerateCert() simply returns those to the caller.
 func TestBothAppAndCATLSAssetsExist(t *testing.T) {
-	ctx := framework.NewTestCtx(t)
+	ctx := framework.NewContext(t)
 	defer ctx.Cleanup()
 	namespace, err := ctx.GetNamespace()
 	if err != nil {
@@ -127,7 +127,7 @@ func TestBothAppAndCATLSAssetsExist(t *testing.T) {
 // it won't verify the existing application TLS cert. Therefore, CertGenerator can't proceed
 // and returns an error to the caller.
 func TestOnlyAppSecretExist(t *testing.T) {
-	ctx := framework.NewTestCtx(t)
+	ctx := framework.NewContext(t)
 	defer ctx.Cleanup()
 	namespace, err := ctx.GetNamespace()
 	if err != nil {
@@ -153,7 +153,7 @@ func TestOnlyAppSecretExist(t *testing.T) {
 // TestOnlyCAExist tests the case where only the CA exists in the cluster;
 // GenerateCert can retrieve the CA and uses it to create a new application secret.
 func TestOnlyCAExist(t *testing.T) {
-	ctx := framework.NewTestCtx(t)
+	ctx := framework.NewContext(t)
 	defer ctx.Cleanup()
 	namespace, err := ctx.GetNamespace()
 	if err != nil {
@@ -182,7 +182,7 @@ func TestOnlyCAExist(t *testing.T) {
 // TestNoneOfCaAndAppSecretExist ensures that when none of the CA and Application TLS assets
 // exist, GenerateCert() creates both and put them into the k8s cluster.
 func TestNoneOfCaAndAppSecretExist(t *testing.T) {
-	ctx := framework.NewTestCtx(t)
+	ctx := framework.NewContext(t)
 	defer ctx.Cleanup()
 	namespace, err := ctx.GetNamespace()
 	if err != nil {
@@ -204,7 +204,7 @@ func TestNoneOfCaAndAppSecretExist(t *testing.T) {
 // TestCustomCA ensures that if a user provides a custom Key and Cert and the CA and Application TLS assets
 // do not exist, the GenerateCert method can use the custom CA to generate the TLS assest.
 func TestCustomCA(t *testing.T) {
-	ctx := framework.NewTestCtx(t)
+	ctx := framework.NewContext(t)
 	defer ctx.Cleanup()
 	namespace, err := ctx.GetNamespace()
 	if err != nil {